refactor: RacingGame클래스에서 출력 기능 분리
출력 기능은 따로 PrintResultUtils로 분리하여 리팩토링
